ZeroMQ as message transport layer (#85)
* uploading benchmark

* skeleton for the zeromq code and tests

* tests passing and major refactor

* new low-level interface

* tests compiling but not yet passing

* first test passing

* most tests passing, response capturing still not implemented

* static message context with overwrites

* req-reply implemented, and all function-call-server tests passing

* all mpi tests passing besides rma + compilation as a library through faasm working as well

* all tests passing

* state tests passing + some refactoring

* updating the function call server/client

* consolidating transport api + refactor

* all tests passing

* cleanup

* adding transport tests

* adding a couple more tests

* switched in-thread REQUIREs to assert-s

* ported flatbuffers and all tests passing

* removing grpc references and bumping version

* cleanup

* removing grpc-root dockerfile

* more cleanup

* changing protoc path

* re-linking zeromq

* updating mpi tests and removing grpc from examples

* linking w/ zeromq

* more verbose tests

* changing install location for zeromq

* updating state as in other PR

* full path in examples

* renaming imported libraries targets

* more cmake work

* adding base container

* fixes for native mpi and removing threaded requires in mpi tests

* pr comments

* more pr comments. introduced message type

* more changes

* refactor tests and move messages to the heap

* freeing memory when possible

* formatting

* fixing teststs

* cleanup

* more detailed error messages + exception checking

* changes after distributed tests

* fix tests

* more test fixing

* state test fixes

* adding send/recv message endpoints

* fix for state tests

* bump version

* minimal breaking test

* out of order tests

* changing ip in the dummy state

* more detailed logging

* using localhost everywhere

* print back trace

* better error reporting

* unwinding nested exceptions

* enforcing unbind and disconnect

* adding monitor for close_event in bind sockets

* cleanup

* more cleanup

* pr comments

+++ src/scheduler/MpiWorld.cpp
@@ -1,6 +1,5 @@
 #include <faabric/mpi/mpi.h>
-#include <faabric/proto/faabric.pb.h>
-#include <faabric/scheduler/FunctionCallClient.h>
+
 #include <faabric/scheduler/MpiThreadPool.h>
 #include <faabric/scheduler/MpiWorld.h>
 #include <faabric/scheduler/Scheduler.h>
@@ -13,6 +12,9 @@
 #include <faabric/util/timing.h>
 
 static thread_local std::unordered_map<int, std::future<void>> futureMap;
+static thread_local std::unordered_map<std::string,
+                                       faabric::scheduler::FunctionCallClient>
+  functionCallClients;
 
 namespace faabric::scheduler {
 MpiWorld::MpiWorld()
@@ -65,6 +67,21 @@ std::shared_ptr<state::StateKeyValue> MpiWorld::getRankHostState(int rank)
     return state.getKV(user, stateKey, MPI_HOST_STATE_LEN);
 }
 
+faabric::scheduler::FunctionCallClient& MpiWorld::getFunctionCallClient(
+  const std::string& otherHost)
+{
+    auto it = functionCallClients.find(otherHost);
+    if (it == functionCallClients.end()) {
+        // The second argument is forwarded to the client's constructor
+        auto _it = functionCallClients.try_emplace(otherHost, otherHost);
+        if (!_it.second) {
+            throw std::runtime_error("Error inserting remote endpoint");
+        }
+        it = _it.first;
+    }
+    return it->second;
+}
+
 int MpiWorld::getMpiThreadPoolSize()
 {
     auto logger = faabric::util::getLogger();
@@ -115,16 +132,48 @@ void MpiWorld::create(const faabric::Message& call, int newId, int newSize)
 
 void MpiWorld::destroy()
 {
-    setUpStateKV();
-    state::getGlobalState().deleteKV(stateKV->user, stateKV->key);
+    // Destroy once per host
+    if (!isDestroyed.test_and_set()) {
+        closeFunctionCallClients();
 
-    for (auto& s : rankHostMap) {
-        const std::shared_ptr<state::StateKeyValue>& rankState =
-          getRankHostState(s.first);
-        state::getGlobalState().deleteKV(rankState->user, rankState->key);
+        // Note - we are deliberately not deleting the KV in the global state
+        // TODO - find a way to do this only from the master client
+
+        for (auto& s : rankHostMap) {
+            const std::shared_ptr<state::StateKeyValue>& rankState =
+              getRankHostState(s.first);
+            state::getGlobalState().deleteKV(rankState->user, rankState->key);
+        }
+
+        localQueueMap.clear();
+    }
+}
+
+void MpiWorld::closeFunctionCallClients()
+{
+    // When shutting down the thread pool, we also make sure we clean all thread
+    // local state by sending a clear message to the queue. Currently, we only
+    // need to close the function call clients
+    for (int i = 0; i < threadPool->size; i++) {
+        std::promise<void> p;
+        threadPool->getMpiReqQueue()->enqueue(
+          std::make_tuple(QUEUE_SHUTDOWN,
+                          std::bind(&MpiWorld::closeThreadLocalClients, this),
+                          std::move(p)));
     }
 
-    localQueueMap.clear();
+    // Lastly clean the main thread as well
+    closeThreadLocalClients();
+}
+
+// Clear thread local state
+void MpiWorld::closeThreadLocalClients()
+{
+    // Close all open sockets
+    for (auto& s : functionCallClients) {
+        s.second.close();
+    }
+    functionCallClients.clear();
 }
 
 void MpiWorld::initialiseFromState(const faabric::Message& msg, int worldId)
@@ -433,10 +482,7 @@ void MpiWorld::send(int sendRank,
         }
     } else {
         logger->trace("MPI - send remote {} -> {}", sendRank, recvRank);
-
-        // TODO - avoid creating a client each time?
-        scheduler::FunctionCallClient client(otherHost);
-        client.sendMPIMessage(m);
+        getFunctionCallClient(otherHost).sendMPIMessage(m);
     }
 }
